container/list: Return as early as possible to avoid unnecessary overhead
For MoveBefore, we can do nothing if e == mark.prev.
In the same way, for MoveAfter, we can do nothing if e == mark.next.

As for other changes, the reason is that 'e.list != l' is unlikely to happen.

Change-Id: I00c6ccd4908f6c14607c1f2449edc090f7c08c22
GitHub-Last-Rev: 786668526c6c8e81317963bc3dfc2feda94820d8
GitHub-Pull-Request: golang/go#62158