time: roll forward with strict RFC 3339 parsing checks
This check was originally reverted in CL 462286 because it
broke a number of projects that implicitly depended on invalid syntax.
This CL rolls forward with the checks to bring the unmarshal behavior
back in compliance with RFC 3339.

To ease the cost of this change for existing systems,
we respect GODEBUG=timerfc3339strict=0 as an override
to opt back into prior behavior.

The runtime/metrics counter for this godebug setting counts
every situation where we detected invalid RFC 3339,
but suppressed the error to maintain backwards compatibility.

Fixes #57912

Change-Id: I920d4f24b10216f270011570ae66dac020f80636