cmd/compile/internal/syntax: add Unparen and UnpackListExpr helpers
We've added Unparen to go/ast, so add syntax.Unparen to be
consistent (and because it's similarly useful).

Also, types2 and noder both have similar functions for unpacking
ListExprs, so might as well add a common implementation in package
syntax too.

Finally, addressing the TODO: UnpackListExpr is small enough to be
inlined, and for typical uses like "range UnpackListExpr(x)" will be
stack allocated. This CL also adds a testing.AllocsPerRun test to
affirm this remains the case.

Change-Id: I96a5591d202193ed5bf1ce6f290919107e3dc01b