cmd/compile/internal/inl: -d=inlinesinglecallfuncs debug flag
DO NOT SUBMIT

Add a new 'InlineSingleCallFuncs' debugging flag that prints out a
report of potential "inline single call" candidate functions. These
are functions that are:

  - not exported
  - not address taken
  - called only once
  - called by a routine that is not an inline candidate

The idea here is that inlines of this sort might be beneficial (even
if over the size limit) in that if we do the inline, then delete the
callee, we've eliminated some call overhead without growing the
program in any meaningful way.

Currently not hooked up to any scoring, e.g. no heuristics taken into
account other than the list above. This patch also doesn't try to
estimate stack frame size effects either, which is potentially
problematic.

Note to self: might also be worth combining this with a more general
dead function elimination phase, to delete not just single call funcs
but zero call funcs as well.

Change-Id: Ia5f968e5201b33f3bbff47a312dd5a34a6b3cf31