make.bat: don't try to download a newer toolchain when calling go version
Calling `go version` will try to download a newer toolchain if the
current one is too old. This is not what we want when building
the toolchain itself. Avoid this by setting `GO111MODULE=off`.

While here, align the unset variables with nogoenv(), present in
make.bash and make.rc.

Fixes #62201

Change-Id: I2d3ac817c2fb161f7a31bf83e32ba934e773b6c2