cmd/compile: retire "IsHiddenClosure" and "IsDeadcodeClosure"
"Hidden" closures used to refer to closures that were declared within
a function, as opposed to closures that appeared in a package-scope
variable initialization expression. This distinction was important
because ir.VisitFuncsBottomUp wants to visit closures within the same
SCC as their enclosing function, so it skipped over hidden closures,
but it needed to still visit the "non-hidden" closures because they
didn't have an enclosing function.

But since CL 522318, all of the package-scope variable initialization
expressions now instead appear in the package's synthetic "init"
function, meaning that closures are now always "hidden".

This CL removes all of the code that worries about hidden vs
non-hidden closures, instead simply looking at fn.OClosure != nil to
distinguish whether a function is a closure or not.

Thanks to Cuong Manh Le for writing the included unit tests, which
he prepared in CL 522956.

Fixes #62277.

Change-Id: Ic3dcd3741d05961f15a6d2f8760fec7dd3cc6ead