net: support canceling Resolver.Lookup[IP/IPAddr/Host/Port] calls on Windows
This CL adds support for canceling DNS lookups on Windows. It does so by
using the [GetAddrInfoEx] function, which supports cancelable async
calls when passing an overlap structure.

Worth noting that the current implementation is still synchronous, as
it uses WaitForSingleObject to wait for GetAddrInfoEx completion. It
would be possible to make it fully asynchronous by plumbing the
overlapped structure with the Go scheduler, but that would require
changes in the netpoll implementation, which at the moment requires
a file descriptor to work, which we don't have one here.

[GetAddrInfoEx]: https://learn.microsoft.com/en-us/windows/win32/api/ws2tcpip/nf-ws2tcpip-getaddrinfoexw

Change-Id: I09329b78810fb3f8e913940920ca8fd8e6d450e5