net: implement host and IP lookups using non-blocking cgo calls on Windows
This CL implements host and IP lookups using GetAddrInfoExW instead of
GetAddrInfoW. GetAddrInfoEx supports asynchronous cancelable calls,
so we can avoid blocking an OS thread while waiting for a DNS response
and also cancel the cgo call if the context is canceled.

Additionally, Resolver.lookupPort has been made cancelable, the context
was not previously being honored.

GetAddrInfoEx returns WSANO_DATA instead of GetAddrInfo's
WSAHOST_NOT_FOUND when the requested name is valid and was found in the
database, but it does not have the correct associated data being
resolved for. When that happens, the returned DNSError will have to
IsNotFound bit set. `TestLookupNoData` is already validating this
behavior. The Unix cgo resolver also maps EAI_NODATA to
IsNotFound, so I don't expect this to be a problem.

Change-Id: I0bf8fa1268f2761211de4015a6e1a26c34a96db0