runtime: give TestFPUnwindAfterRecovery more opportunities to fail
Currently TestFPUnwindAfterRecovery tries to invoke a crash, but that
crash isn't deterministic. In fact, when backporting the fix for the
regression the test is checking for, I couldn't get the test to fail
without just verifying the frames that come out of unwinding.

So, validate the frames that come out of unwinding. This test update
will be backported.

For #61766.

Change-Id: I2ae509b4ed71f5c759bca6b731a7961f03a0bde7