log/slog: fix Record.back slice always to small during Add()
When `slog.Record.Add(args)` is called, with enough args to cause the `Record.back` `[]Attr` to be created, it is being created 1 too small, which results in it immediately being grown again by `append` before the function exits (needless allocation and copying).  This is because it is created with a capacity equal to `countAttrs(args)`, but forgets that there is an additional attribute to be appended: `a` (`args` is just the remaining unconsumed attributes).  This PR fixes that by adding 1 to the capacity to account for the `a` attribute.
```go
if r.back == nil {
	r.back = make([]Attr, 0, countAttrs(args)) // args is the unconsumed attributes not counting 'a'
}
r.back = append(r.back, a) // on the last round of the for loop, this will grow the slice again
```

Additionally, when `Record.back` already exists, it will most likely be at max capacity already.  Rather than append to it and risk having it grown multiple times, or grow too large, I've added a `slices.Grow` call to set it to the right capacity, similar to what is already done in the `Record.AddAttrs(attrs)` function.

Change-Id: Ic4bcf45909fe4436c586ccd2b8d61f24606b6be8
GitHub-Last-Rev: 22d702fc7b84778c14072a13bc24b4f5d309709f
GitHub-Pull-Request: golang/go#62388