cmd/compile/internal/escape: reserve space in leaks encoding for interface receiver data
WIP. Feedback welcome.

In preparation for analyzing receivers in interface method calls,
update the leaks encoding.

The leaks zeroth element still communicates information about
flows to the heap, but the next element now communicates
information about use as an interface receiver, which might
or might not eventually flow to the heap.

While doing so, we also change the magic prefix to purposefully
break any readers that we do not know about. A future CL might change
it back to the prior esc: prefix.

TODO: consider keeping the esc: prefix unchanged in this CL.
Change-Id: Ie3131f41466cdd6df8e6a30399ac5947e5ea4ee1