cmd/compile/internal/escape: analyze receivers in interface method calls
WIP. Feedback welcome.

This is part of a series of CLs that aim to reduce how often
interface arguments escape for the print functions in fmt.

Here, we update escape analysis to examine interface method calls
and attempt to prove that the value used as the interface receiver
argument cannot be leaked by the implementing method given what
we know about the value's type at the call site.

We start with some background. Consider the following example.
Currently, the use of the input parameter as a receiver in an
interface method call causes the current escape analysis to
record that the input parameter to this function always
leaks to the heap:

  func Print(input any) {
    if v, ok := input.(Stringer); ok {
       println(v.String())
    }
  }

Escape analysis determines this by first walking the AST and building up
a data-flow graph. For this example, the nodes in the graph
are the three locations input, v, and <heap>, a pseudo location
in the data-flow graph representing that something must be stored in
the heap. The edges of the graph indicate the data flow from location
to location. This CL adds a simple dump of the data-flow graph, which
for this example looks like:

  location 'input' in Print
     incoming edges: none
  location 'v' in Print
     incoming edge from 'input' in Print
  location '<heap>'
     incoming edge from 'v' in Print

The edges showing the flow from input to v, and then from v to <heap>
(because the AST has v being used as the receiver in an interface
method call), but escape analysis does not yet know if there is
connectivity between input and <heap>. To determine that,
escape analysis runs a shortest path algorithm in solve.go to
conclude that input does indeed have a path to <heap>, and hence
input is recorded as escaping.

When Print is later used, the argument allocates:

  val := 1000
  Print(val)  // val escapes to heap

The data-flow graph is a simplified model that ignores conditionals,
does not generally distinguish between elements within a compound
variable, and so on. Even if our Print example is called with an
int value, it still currently escapes, even though an int has no
methods, will never succeed in the type assertion, and will
never execute the String method call.

To improve this, this CL introduces an additional pseudo location
in the data-flow graph.

The data-flow graph after walking the AST for our example is the
same as before, except instead of a <heap> pseudo location, the
example graph now has an <interface receiver> pseudo location
to indicate v was used as the receiver in an interface method call:

  location 'input' in Print
     incoming edges: none
  location 'v' in Print
     incoming edge from 'input' in Print
  location '<interface receiver>'
     incoming edge from 'v' in Print

The shortest path algorithm sees that input has a path to the
<interface receiver> in Print, but that is inconclusive -- we
don't yet know what type will be passed to Print. We therefore
record a new parameter tag type on the input argument
that allows us to propagate this uncertainty across function call
boundaries.

Now, when Print is later used, escape analysis has been updated
to analyze the type at the call site. Here, it is able to prove
val cannot be leaked solely based on its use as an interface receiver
in Print, and hence records that the argument no longer allocates:

  val := 1000
  Print(val)  // val does not escape

The type-specific analysis is conservative (e.g., currently
gives up on recursive types), but can make use of recorded
escape analysis results for the methods of the call site type.
It does not currently pass method-specific information in the
data-flow graph, so a single escaping receiver on one method
for type Foo means Foo will never be proven as non-escaping
for any interface method call (though handling that could
be an extension in the future).

For arguments to the print family in fmt, this change
means many types no longer allocate due to fmt checking the
arguments for Stringer, GoStringer, Formatter, and so on.

However, the arguments still escape for other reflect-related
reasons, which are hopefully the topic of the next set of CLs.

The tests for this CL include:

  escape_iface_recv_basic.go -- An overview and good place to start.
  escape_iface_recv_extracted.go -- Simplified examples extracted
      from std and cmd, and which starts with a simplified fmt.Print.
  escape_iface_recv.go -- More variations.
  escape_basic.go -- Not strictly related to this work aside from
      serving as a resource for sanity checking how pointers vs.
      interfaces are handled by escape analysis in general.

Updates #8618

Change-Id: I76c30a9de65865fb407865776ddc258b3f859e2c