cmd/link/internal/ld: change dwctxt methods to use pointer receivers
A bunch of these methods were using value receivers, which is a bit
weird for such a large type that is used to implement an interface.

Also, directly embed the sync.Mutex, rather than indirectly.
(Presumably it was originally a pointer to workaround many of the
methods being value receivers?)

Change-Id: I925c5771bf1c9a614520237ed789b3cd91ed1e84