encoding/json: zstd compress the testdata
There is a non-public zstd decoder in the stdlib (CL 473356).

Using zstd v1.4.4, golang_source.json can be recompressed
with `zstd -19` to become ~91.502KiB large instead of the gzip'd
size of ~117.609KiB. This is a reduction of 22%.

This commit will increase the Git history by ~92KiB,
but reduce the final Go release by ~26KiB.
Presumably there are orders and orders of magnitude more downloads
of the final release than there are checkouts of the Git repo.

This commit will set the precedence for using zstd for other
testdata in the stdlib. For example, the largest testdata directory
is internal/trace/testdata at ~2.218MiB. With gzip compression
using zopfli, we can reduce this to ~1.212MiB.
However, with zstd, we can reduce this to ~1.101MiB.
Another compression format with better ratios is XZ,
which results in a size of ~0.951MiB.

However, there is no native XZ decompressor in the stdlib,
the LZMA algorithm is not well specified, and
LZMA decompression is really slow.
Zstandard probably represents the best balance of tradeoffs
for the next decade to come. It is well-specified,
fast to decompress, and an implementation exists in stdlib.

Change-Id: I6da2df27bd260befc0a44c6bc0255365be0a5b0f