cmd/compile/internal/escape: handle interface method calls better
DO NOT SUBMIT: Needs tests.

DO NOT SUBMIT: Needs a solution to #62501. In particular, the changes
in src/net and to test/escape_reflect.go should not be necessary.

Currently, if a value of concrete type T is converted to interface
type, and then that interface value has any methods called on it, then
we conservatively assume the T value must flow to the heap.

But we can actually be smarter about this: we can instead recognize
that the T will always flow to one of the receiver parameters in its
method set. This is thanks to an idea from thepudds, who prototyped it
at go.dev/cl/524945.

This CL is a separate attempt at the same idea.

In particular, if we notice a T->I conversion result flows to the
receiver parameter of an interface method call, then we flow the T
value to all of its receiver parameters.

To make this work, we also need the call dependency graph used by
escape analysis to treat any T->I conversion as implicitly
dependending on all of T's methods.

Unfortunately, adding these additional call graph edges ends up
creating more cycles, which can in turn cause functions to be
analyzed *less* precisely due to #62501.

Change-Id: If4711fee3dca2cf668b664f244eb909dc34f3d28