testing: add Chdir
DO NOT REVIEW
DO NOT MERGE
(this is a WIP, squashed to test the changes)

Some tests need to use os.Chdir, but the use is complicated because
 - old PWD value needs to be set after the test is done;
 - such tests can not be run in Parallel.

Add Chdir that covers these cases.

While at it, unify the panic message when t.Parallel is used together
with t.Setenv or t.Chdir, and fix the test case accordingly.

Fixes #62516

Change-Id: Ib116ec47fbc20ed503a1a42869465b210d1923be